Refactor report building architecture
